Restructured deid table files into multiple files
